services/worker/pyproject.toml
[project]
name = "tkp-worker"
version = "0.1.0"
description = "Ingestion worker service for multi-tenant knowledge platform"
requires-python = ">=3.12"
dependencies = [
    "minio>=7.2.18",
    "oss2>=2.19.1",
    "pydantic-settings>=2.11.0",
    "psycopg[binary]>=3.2.9",
    "sqlalchemy>=2.0.43",
    "openai>=1.54.0",
    "pgvector>=0.4.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.2",
    "pillow>=10.0.0",
    "tiktoken>=0.7.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.3",
    "camelot-py[cv]>=0.11.0",
    "tabula-py>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tkp_worker"]
